requirements.txt
spacy==3.7.4
emoji==2.11.1
orjson==3.10.3
pandas==2.1.4
numpy==1.24.4
tqdm==4.66.1
pyyaml==6.0.1
python-dateutil==2.8.2